# Conflicts:
#	experience/src/main/java/com/senai/experience/controllers/ItemPedidoController.java
#	experience/src/main/java/com/senai/experience/controllers/PedidoController.java
#	experience/src/main/java/com/senai/experience/controllers/VeiculoController.java
#	experience/src/main/java/com/senai/experience/entities/Pedido.java
#	experience/src/main/java/com/senai/experience/services/PedidoService.java

+++ experience/src/main/java/com/senai/experience/DTO/request/VeiculoRequest.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class VeiculoRequest {
-    private int idProduto;
+    private Long idProduto;
     private int chassi;
     private String statusVeiculo;
 }
